+++ validator/impl/collator.cpp
@@ -811,7 +811,7 @@ bool Collator::request_out_msg_queue_size() {
   for (size_t i = 0; i < prev_blocks.size(); ++i) {
     ++pending;
     send_closure_later(manager, &ValidatorManager::get_out_msg_queue_size, prev_blocks[i],
-                       [self = get_self(), i](td::Result<td::uint32> res) {
+                       [self = get_self(), i](td::Result<td::uint64> res) {
                          td::actor::send_closure(std::move(self), &Collator::got_out_queue_size, i, std::move(res));
                        });
   }
@@ -890,14 +890,14 @@ void Collator::got_neighbor_out_queue(int i, td::Result<Ref<MessageQueue>> res) 
  * @param i The index of the previous block (0 or 1).
  * @param res The result object containing the size of the queue.
  */
-void Collator::got_out_queue_size(size_t i, td::Result<td::uint32> res) {
+void Collator::got_out_queue_size(size_t i, td::Result<td::uint64> res) {
   --pending;
   if (res.is_error()) {
     fatal_error(
         res.move_as_error_prefix(PSTRING() << "failed to get message queue size from prev block #" << i << ": "));
     return;
   }
-  td::uint32 size = res.move_as_ok();
+  td::uint64 size = res.move_as_ok();
   LOG(WARNING) << "got outbound queue size from prev block #" << i << ": " << size;
   out_msg_queue_size_ += size;
   check_pending();
